Combine amount and count time series into one `transactionsTimeSeries` returning transactionAmountWithCount nodes
