+++ server/graphql/v2/collection/AccountCollection.js
@@ -1,7 +1,5 @@
-import { GraphQLBoolean, GraphQLList, GraphQLNonNull, GraphQLObjectType } from 'graphql';
-import { Op } from 'sequelize';
+import { GraphQLList, GraphQLNonNull, GraphQLObjectType } from 'graphql';
 
-import models from '../../../models';
 import { Account } from '../interface/Account';
 import { Collection, CollectionFields } from '../interface/Collection';
 import { AccountCollectionStats } from '../object/AccountCollectionStats';
@@ -18,21 +16,9 @@ const AccountCollection = new GraphQLObjectType({
       },
       stats: {
         type: new GraphQLNonNull(AccountCollectionStats),
-        args: {
-          includeChildren: { type: GraphQLBoolean, defaultValue: false },
-        },
         description: 'Stats for the returned results (i.e. accounts within the limit)',
-        async resolve(collection, args) {
-          const collectiveIds = collection.nodes.map(c => c.id);
-
-          if (args.includeChildren) {
-            const childCollectives = await models.Collective.findAll({
-              attributes: ['id'],
-              where: { ParentCollectiveId: { [Op.in]: collectiveIds } },
-            });
-            collectiveIds.push(...childCollectives.map(c => c.id));
-          }
-          return { collectiveIds };
+        resolve(collection) {
+          return collection;
         },
       },
     };
