+++ server/graphql/v2/object/AccountCollectionStats.ts
@@ -1,19 +1,28 @@
-import { GraphQLList, GraphQLNonNull, GraphQLObjectType } from 'graphql';
+import { GraphQLBoolean, GraphQLList, GraphQLNonNull, GraphQLObjectType } from 'graphql';
 import moment from 'moment';
+import { Op } from 'sequelize';
 
 import queries from '../../../lib/queries';
+import models from '../../../models';
 import { TransactionKind } from '../enum/TransactionKind';
 import { TransactionType } from '../enum/TransactionType';
-import { resultsToAmountNode } from '../object/HostMetricsTimeSeries';
-import { getNumberOfDays, getTimeUnit, TimeSeriesAmount, TimeSeriesArgs } from '../object/TimeSeriesAmount';
-import { TimeSeriesCount } from '../object/TimeSeriesCount';
+import { getNumberOfDays, getTimeUnit, TimeSeriesArgs } from '../object/TimeSeriesAmount';
+import { TimeSeriesAmountWithCount } from '../object/TimeSeriesAmountWithCount';
+
+export const resultsToAmountWithCountNode = results => {
+  return results.map(result => ({
+    date: result.date,
+    count: result.count,
+    amount: { value: result.amount, currency: result.currency },
+  }));
+};
 
 export const AccountCollectionStats = new GraphQLObjectType({
   name: 'AccountCollectionStats',
   description: 'Account collection stats',
   fields: () => ({
-    transactionsCountTimeSeries: {
-      type: new GraphQLNonNull(TimeSeriesCount),
+    transactionsTimeSeries: {
+      type: new GraphQLNonNull(TimeSeriesAmountWithCount),
       args: {
         ...TimeSeriesArgs,
         kind: {
@@ -24,35 +33,20 @@ export const AccountCollectionStats = new GraphQLObjectType({
           type: TransactionType,
           description: 'The transaction type (DEBIT or CREDIT)',
         },
+        includeChildren: { type: GraphQLBoolean, defaultValue: false },
       },
-      description: 'Time series of the number of transactions to these accounts',
-      resolve: async ({ collectiveIds }, args) => {
-        const dateFrom = args.dateFrom ? moment(args.dateFrom) : null;
-        const dateTo = args.dateTo ? moment(args.dateTo) : null;
-        const timeUnit = args.timeUnit || getTimeUnit(getNumberOfDays(dateFrom, dateTo, {}) || 1);
+      description: 'Time series of the transaction count and sum of amount for these accounts',
+      resolve: async (collection, args) => {
+        const collectiveIds = collection.nodes.map(c => c.id);
 
-        const transactionParams = { type: args.type, kind: args.kind, dateFrom, dateTo, collectiveIds };
-        const countNodes = collectiveIds.length
-          ? await queries.getTransactionsCountTimeSeries(timeUnit, transactionParams)
-          : [];
-        return { dateFrom, dateTo, timeUnit, nodes: countNodes };
-      },
-    },
-    transactionsAmountTimeSeries: {
-      type: new GraphQLNonNull(TimeSeriesAmount),
-      args: {
-        ...TimeSeriesArgs,
-        kind: {
-          type: new GraphQLList(TransactionKind),
-          description: 'To filter by transaction kind',
-        },
-        type: {
-          type: TransactionType,
-          description: 'The transaction type (DEBIT or CREDIT)',
-        },
-      },
-      description: 'Time series of the sum of transactions to these accounts',
-      resolve: async ({ collectiveIds }, args) => {
+        if (args.includeChildren) {
+          const childCollectives = await models.Collective.findAll({
+            attributes: ['id'],
+            where: { ParentCollectiveId: { [Op.in]: collectiveIds } },
+          });
+          collectiveIds.push(...childCollectives.map(c => c.id));
+        }
+
         const dateFrom = args.dateFrom ? moment(args.dateFrom) : null;
         const dateTo = args.dateTo ? moment(args.dateTo) : null;
         const timeUnit = args.timeUnit || getTimeUnit(getNumberOfDays(dateFrom, dateTo, {}) || 1);
@@ -61,7 +55,7 @@ export const AccountCollectionStats = new GraphQLObjectType({
         const amountDataPoints = collectiveIds.length
           ? await queries.getTransactionsTimeSeries(timeUnit, transactionParams)
           : [];
-        return { dateFrom, dateTo, timeUnit, nodes: resultsToAmountNode(amountDataPoints) };
+        return { dateFrom, dateTo, timeUnit, nodes: resultsToAmountWithCountNode(amountDataPoints) };
       },
     },
   }),
