+++ server/lib/queries.js
@@ -1134,7 +1134,7 @@ const getTransactionsTimeSeries = async (
   { collectiveIds = [], type = null, kind = null, dateFrom = null, dateTo = null } = {},
 ) => {
   return sequelize.query(
-    `SELECT DATE_TRUNC(:timeUnit, "createdAt") AS "date", sum("amountInHostCurrency") as "amount", "hostCurrency" as "currency"
+    `SELECT DATE_TRUNC(:timeUnit, "createdAt") AS "date", sum("amountInHostCurrency") as "amount", "hostCurrency" as "currency", count("id") as "count"
          FROM "Transactions"
          WHERE "deletedAt" IS NULL
            AND "CollectiveId" IN (:collectiveIds)
@@ -1143,38 +1143,6 @@ const getTransactionsTimeSeries = async (
            ${dateFrom ? `AND "createdAt" >= :startDate` : ``}
            ${dateTo ? `AND "createdAt" <= :endDate` : ``}
          GROUP BY DATE_TRUNC(:timeUnit, "createdAt"), "hostCurrency"
-         ORDER BY DATE_TRUNC(:timeUnit, "createdAt")
-        `,
-    {
-      type: sequelize.QueryTypes.SELECT,
-      replacements: {
-        kind: Array.isArray(kind) ? kind : [kind],
-        type,
-        timeUnit,
-        collectiveIds,
-        ...computeDatesAsISOStrings(dateFrom, dateTo),
-      },
-    },
-  );
-};
-
-/**
- * Returns the count of transactions over time.
- */
-const getTransactionsCountTimeSeries = async (
-  timeUnit,
-  { collectiveIds = [], type = null, kind = null, dateFrom = null, dateTo = null } = {},
-) => {
-  return sequelize.query(
-    `SELECT DATE_TRUNC(:timeUnit, "createdAt") AS "date", count("id") as "count"
-         FROM "Transactions"
-         WHERE "deletedAt" IS NULL
-           AND "CollectiveId" IN (:collectiveIds)
-           ${type ? `AND "type" = :type` : ``}
-           ${kind?.length ? `AND "kind" IN (:kind)` : ``}
-           ${dateFrom ? `AND "createdAt" >= :startDate` : ``}
-           ${dateTo ? `AND "createdAt" <= :endDate` : ``}
-         GROUP BY DATE_TRUNC(:timeUnit, "createdAt")
          ORDER BY DATE_TRUNC(:timeUnit, "createdAt")
         `,
     {
@@ -1234,7 +1202,6 @@ const queries = {
   getUniqueCollectiveTags,
   getGiftCardBatchesForCollective,
   getTransactionsTimeSeries,
-  getTransactionsCountTimeSeries,
 };
 
 export default queries;
